requirements.txt
asgiref==3.4.1
bleach==4.1.0
Django==3.2.9
django-colorfield==0.6.0
django-grappelli==2.15.3
django-nested-admin==3.3.3
django-phonenumber-field==6.0.0
django-summernote==0.8.20.0
gunicorn==20.1.0
mysqlclient==2.1.0
numpy==1.21.4
packaging==21.2
phonenumbers==8.12.39
Pillow==8.4.0
pyparsing==2.4.7
python-monkey-business==1.0.0
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
ua-parser==0.10.0
user-agents==2.2.0
webencodings==0.5.1
